Exercise 163:Words with Six Vowels in Order
There is at least one word in the English language that contains each of the vowels
a, e, i, o, u and y exactly once and in order. Write a program that searches a file
containing a list of words and displays all of the words that meet this constraint. The
user will provide the name of the file that will be searched. Display an appropriate
error message and exit the program if the user provides an invalid file name or if
something else goes wrong while searching for words with six vowels in order.