﻿ Exercise 172: Element Sequences
Another game that some people play with the names of chemical elements involves
constructing a sequence of elements where each element in the sequence begins with
the last letter of its predecessor. For example, if a sequence begins with Hydrogen,
then the next element must be an element that begins with N, such as Nickel. The
element following Nickel must begin with L, such as Lithium. The element sequence
that is constructed can not contain any duplicates.
Write a program that reads the name of an element from the user. Your program
should use a recursive function to find the longest sequence of elements that begins
with the entered element. Then it should display the sequence. Ensure that your
program responds in a reasonable way if the user does not enter a valid element name.
Hint: It may take your program up to two minutes to find the longest sequence
for some elements. As a result, you might want to use elements like Molybdenum and Magnesium as your first test cases. Each has a longest sequence
that is only 8 elements long which your program should find in a fraction of a
second.